requirements.txt
streamlit>=1.22.0
openai>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
pandas>=1.3.0
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6
requests>=2.31.0
protobuf-to-pydantic>=0.2.5
openai-agents>=1.0.0
